REF: set implementation
